R STUDIO LANGUAGE PREACHER AND HAYES TEST
Hello everyone, 
im doing a medical research and to finish my project i need to make a preacher and hayes test on r studio. I search on the internet but i did not find the script. Does anyone know what package and fonctions are needed to realise the preacher and hayes test on r studio? 
Thanks everyone